fix: close RCA audit findings across core, app, GPU, services, CI, and Jenkins (#71)
Full codebase RCA audit: 6 issues created, verified, and resolved.

Code fixes:
- fix(app): clamp Normal mouse coordinates to X10 protocol maximum (222)
  to prevent silent u16-to-u8 truncation on wide terminals (closes #65)
- fix(core): reject out-of-range SGR 256-color indices and RGB components
  via u8::try_from instead of silent truncation (closes #66)
- fix(gpu): cast operands to u64 before multiplication in all 5 GPU buffer
  sizing paths to prevent usize overflow on 32-bit targets (closes #67)
- fix(services): remove dead RenderModeController::set_mode wrapper that
  silently discarded #[must_use] transition result (closes #68)

CI/CD fixes:
- ci: add version comments to all SHA-pinned actions across all workflows
  and all Dockerfile base images for consistent auditability (closes #69)
- fix(jenkins): remove duplicate disableConcurrentBuilds in JCasC pipeline,
  restoring intended abortPrevious semantics (closes #70)
- fix(jenkins): pass CARGO_TARGET_DIR as env var instead of inline in
  codeql --command, fixing persistent server-side CodeQL failures

Tests: 890 total (885 + 5 regression), 0 failures.
Reviewed by: code-reviewer, silent-failure-hunter, Codex.

+++ crates/features/render_gpu/src/bootstrap.rs
@@ -264,7 +264,7 @@ impl GpuRenderer {
         });
 
         let initial_capacity = initial_cell_buffer_capacity(config.width, config.height);
-        let cell_buf_size = (initial_capacity * std::mem::size_of::<CellInstance>()) as u64;
+        let cell_buf_size = initial_capacity as u64 * std::mem::size_of::<CellInstance>() as u64;
         let cell_buf_usage = wgpu::BufferUsages::STORAGE
             | wgpu::BufferUsages::COPY_DST
             | wgpu::BufferUsages::COPY_SRC;
